+++ ZeroLibraries/SpatialPartition/SapContainers.hpp
@@ -35,8 +35,8 @@ public:
   {
     //get the id's of the two client data's
     HashPolicy<ClientDataType> hasher;
-    uint id1 = hasher(data1.mClientData);
-    uint id2 = hasher(data2.mClientData);
+    uint id1 = (uint)hasher(data1.mClientData);
+    uint id2 = (uint)hasher(data2.mClientData);
     //make sure the aren't the same
     if(id1 == id2)
       return;
@@ -66,8 +66,8 @@ public:
   {
     //get the id's of the two client data's
     HashPolicy<ClientDataType> hasher;
-    uint id1 = hasher(data1.mClientData);
-    uint id2 = hasher(data2.mClientData);
+    uint id1 = (uint)hasher(data1.mClientData);
+    uint id2 = (uint)hasher(data2.mClientData);
     //get the lexicographic index of this pair
     PairId index = GetLexicographicId(id1,id2);
 
@@ -270,7 +270,7 @@ struct SapRange
   {
     //we have to loop over all of the boxes because we can't otherwise
     //deal with a box that completely Contains the query aabb
-    uint end = mBoxes->Size();
+    uint end = (uint)mBoxes->Size();
     for(; mIndex < end; ++mIndex)
     {
       BoxType& box = (*mBoxes)[mIndex];
@@ -332,7 +332,7 @@ struct SapPairRange
 
   uint Length()
   {
-    return mRange.Size();
+    return (uint)mRange.Size();
   }
 
   typename PairManagerType::range mRange;
